requirements.txt
Flask>=2.3,<4.0
Flask-SQLAlchemy>=3.0,<4.0
Flask-JWT-Extended>=4.0,<5.0
Flask-CORS>=3.0,<5.0
SQLAlchemy>=2.0,<3.0
Werkzeug>=2.3,<4.0
marshmallow>=3.19,<4.0
python-dotenv>=0.21,<2.0
requests>=2.28,<3.0
numpy>=1.25,<2.0
pandas>=2.0,<3.0
scikit-learn>=1.3,<2.0
textblob>=0.17,<0.18
vaderSentiment>=3.3,<4.0
googlemaps>=4.9,<5.0
geopy>=2.4,<3.0
folium>=0.14,<0.16
Flask-Migrate>=3.0,<4.0